requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
numpy==1.26.4
faiss-cpu==1.8.0.post1
fastembed==0.3.3
onnxruntime==1.18.1
langdetect==1.0.9
twilio==9.0.5
httpx==0.27.0
python-dotenv==1.0.1
orjson==3.10.3
pytz==2024.1
openai>=1.30.0
python-docx==0.8.11
pdfminer.six==20240706
